PycharmProjects/Voice Search/requirements.txt
fastapi>=0.95,<1.0
uvicorn[standard]>=0.22,<0.23
python-dotenv>=1.0,<2.0
pydantic>=2.5,<3.0
SpeechRecognition>=3.14,<4.0
pyaudio>=0.2.14,<1.0
google-cloud-translate>=3.21,<4.0
pydantic-settings>=0.3,<1.0